refactor(forges/github): relocate GitHub modules into forges/github
