+++ src/renderer/context/App.tsx
@@ -37,7 +37,6 @@ import type {
   LoginPersonalAccessTokenOptions,
 } from '../utils/auth/types';
 
-import { clearOctokitClientCache } from '../utils/api/octokit';
 import {
   exchangeAuthCodeForAccessToken,
   performGitHubWebOAuth,
@@ -52,6 +51,7 @@ import {
   removeAccount,
 } from '../utils/auth/utils';
 import { clearState, loadState, saveState } from '../utils/core/storage';
+import { clearOctokitClientCache } from '../utils/forges/github/octokit';
 import { getAdapterById } from '../utils/forges/registry';
 import {
   applyKeyboardShortcut,
