+++ src/renderer/utils/forges/github/graphql/MergeQueryBuilder.ts
