+++ src/renderer/utils/forges/github/graphql/common.graphql
