+++ src/renderer/utils/forges/github/graphql/discussion.graphql
