+++ src/renderer/utils/forges/github/graphql/merged.graphql
